Updating project_detail page with latest changes so that we can work on making project detail page dynamic

+++ laravel-backend/app/Http/Controllers/Api/v1/ProjectController.php
@@ -33,8 +33,24 @@ class ProjectController extends Controller
                 "message" => "You're not authorized to perform this action"
             ], 403);
         }
-        return new ProjectResource(Project::create($request->all()));
+        $data = $request->all();
+        $user = $request->user();
+        $data['team_lead'] = $user->name;
 
+        // Create new project, with project_lead attribute set to user(professor)'s name
+        $project = new ProjectResource(Project::create($data));
+        
+        // Create a new project_membership with role as 'lead' and status as 'active'
+        $user->projects()->attach($project->id, [
+            'role' => 'lead',
+            'status' => 'active',
+            'user_email' => $user->email
+        ]);
+
+        return response()->json([
+            'message' => "Project created successfully and $user->name is the lead for projec",
+            'data' => $project
+        ]);
     }
 
     /**
@@ -50,8 +66,9 @@ class ProjectController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateProjectRequest $request, Project $project)
-    {
+    public function update(UpdateProjectRequest $request, $id)
+    {   
+
         // A user with role == Student cannot update project
         $role = $request->user()->role;
         if ($role == "student" || $role == "Student"){
@@ -60,14 +77,24 @@ class ProjectController extends Controller
             ], 403);
         }
 
-        $project->update($request->all());
+        $project = Project::find($id);
+        
+        if(!$project){
+            return response()->json([
+                "message" => "resource not found"]
+            , 404);
+        }
+
+
+        $validated = $request->validated();
+        $project->update($validated);
         return new ProjectResource($project);
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Project $project)
+    public function destroy($id)
     {
         // A user with role == Student cannot delete a project
         $role = auth()->user()->role;
@@ -77,7 +104,16 @@ class ProjectController extends Controller
             ], 403);
         }
 
+        $project = Project::find($id);
+        
+        if(!$project){
+            return response()->json([
+                "message" => "resource not found"]
+            , 404);
+        }
+
         $project->delete();
-        return response()->json(['message'=>"$project->title Deleted Successfully"]);
+        return response()->json(
+            ['message'=>"$project->title Deleted Successfully"]);
     }
 }
